+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Python: Added ZRANDMEMBER command ([#1413](https://github.com/aws/glide-for-redis/pull/1413))
 * Python: Added BZMPOP command ([#1412](https://github.com/aws/glide-for-redis/pull/1412))
 * Python: Added ZINTERCARD command ([#1418](https://github.com/aws/glide-for-redis/pull/1418))
+* Python: Added ZMPOP command ([#1417](https://github.com/aws/glide-for-redis/pull/1417))
 
 
 #### Fixes